Refactored UserPrompt into Modal
