+++ src/ui/components/shared/Error.css
@@ -1,0 +1,34 @@
+.modal-container.error .modal {
+  text-align: center;  
+}
+
+.modal-container.error .modal button.clicked {
+  background: linear-gradient(152.34deg, #3398ff 25.66%, #007dfd 78.07%);
+}
+
+.modal-container.error .modal button.clicked .img.refresh {
+  animation: spin .7s linear infinite;
+}
+
+.modal-container.error .modal .tip {
+  color: var(--light-grey);
+  width: 75%;
+  font-size: 1.1rem;
+  line-height: 1.5;
+  background-image: "/images"
+}
+
+.modal-container.error .modal .tip .record {
+  margin: 0 .5rem;
+  border: 1px solid var(--faint-red);
+  padding: 2px 5px;
+  border-radius: 5px;
+  font-size: .8rem;
+  vertical-align: text-bottom;
+  color: var(--faint-red)
+}
+
+.modal-container.error .modal button .img.refresh {
+  margin-inline-end: .5em;
+  background-color: white;
+}
