+++ src/ui/components/shared/Error.js
@@ -1,7 +1,9 @@
 import React, { useState } from "react";
-import UserPrompt from "./UserPrompt";
+import Modal from "./Modal";
 import classnames from "classnames";
 import { useAuth0 } from "@auth0/auth0-react";
+
+import "./Error.css";
 
 function RefreshButton() {
   const [clicked, setClicked] = useState(false);
@@ -20,22 +22,22 @@ function RefreshButton() {
 
 export function PopupBlockedError() {
   return (
-    <UserPrompt classnames={["error", "overlay"]}>
+    <Modal opaque error>
       <h1>Uh-oh</h1>
       <p>Please turn off your pop-up blocker and refresh this page.</p>
       <RefreshButton />
-    </UserPrompt>
+    </Modal>
   );
 }
 
 export function SessionError({ error }) {
   return (
-    <UserPrompt classnames={["error", "overlay"]}>
+    <Modal translucent error>
       <h1>Whoops!</h1>
       <p>Looks like something went wrong with this page.</p>
       <RefreshButton />
       <p className="tip">Error: {error}</p>
-    </UserPrompt>
+    </Modal>
   );
 }
 
@@ -43,7 +45,7 @@ export function UnauthorizedAccessError() {
   const { loginWithRedirect, isAuthenticated } = useAuth0();
 
   return (
-    <UserPrompt classnames={["error"]}>
+    <Modal opaque error>
       <h1>This is a private recording</h1>
       {!isAuthenticated ? (
         <>
@@ -53,6 +55,6 @@ export function UnauthorizedAccessError() {
       ) : (
         <p>You don&apos;t have permission to view this recording.</p>
       )}
-    </UserPrompt>
+    </Modal>
   );
 }
