+++ src/ui/components/shared/Sharing.css
@@ -1,0 +1,104 @@
+/* Sharing modal styles */
+
+.modal-container .permissions-list {
+  width: 40rem;
+  display: flex;
+  flex-direction: column;
+}
+
+.modal-container .permission {
+  display: flex;
+  align-items: center;
+  padding: 12px 4px;
+  transition-duration: 200ms;
+}
+
+.modal-container .permission:hover {
+  background: #EEE;
+  border-radius: 8px;
+}
+
+.modal-container .permission:not(:last-child) {
+  margin-bottom: 4px;
+}
+
+.modal-container .permission .icon {
+  height: 3rem;
+  width: 3rem;
+  border-radius: 50%;
+  margin: 0 12px 0 8px;
+  flex-shrink: 0;
+  background: var(--dark-grey);
+  background-position: center;
+  background-size: contain;
+}
+
+.modal-container .permission .main {
+  flex-grow: 1;
+}
+
+.modal-container .permission .role {
+  flex-shrink: 0;
+  padding: 0px 8px;
+  color: var(--light-grey)
+}
+
+.modal-container .modal form {
+  display: flex;
+  flex-direction: row;
+}
+
+.modal-container .modal input[type=text] {
+  height: 40px;
+  display: block;
+  flex-grow: 1;
+  padding: 1rem;
+  text-align: center;
+  font-size: 1.25rem;
+  margin: auto;
+  border: 1px solid var(--light-grey);
+  color: var(--light-grey);
+}
+
+.modal-container .modal input[type=submit] {
+  display: none;
+  flex-shrink: 0;
+  color: white;
+  background: var(--dark-blue);
+  border: 0px;
+  padding: 0px 2rem;
+  border-radius: 1rem;
+}
+
+.modal-container .modal form:focus-within input[type=submit] {
+  display: block;
+  margin-left: 1rem;
+}
+
+.modal-container .buttons {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+}
+
+.modal-container .spinner {
+  display: flex;
+  height: 40px;
+  flex-direction: row;
+  justify-content: center;
+  flex-grow: 1;
+  padding: 1rem;
+  text-align: center;
+  font-size: 1.25rem;
+  margin: auto;
+  border: 1px solid var(--light-grey);
+  color: var(--light-grey);
+}
+
+.modal-container .spinner .img.refresh {
+  height: 16px;
+  padding-inline-end: 1em;
+  margin-inline-end: .5em;
+  background-color: var(--light-grey);
+  animation: spin .7s linear infinite;
+}
