--- src/ui/components/shared/UserPrompt.css
@@ -1,130 +1,0 @@
-.user-prompt {
-  display: flex;
-  flex-direction: column;
-  padding: 5rem;
-  margin: auto;
-  align-items: center;
-  font-size: 1.25rem;
-  text-align: center;
-  width: 630px;
-}
-
-.user-prompt p, .user-prompt li {
-  color: var(--dark-grey);
-  line-height: 1.5;
-}
-
-.user-prompt ol {
-  text-align: left;
-}
-
-.user-prompt > *:not(:last-child) {
-  margin-bottom: 2rem;
-}
-
-.user-prompt form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.user-prompt input {
-  display: block;
-  width: 40rem;
-  padding: 1rem;
-  text-align: center;
-  font-size: 1.25rem;
-  margin: auto;
-  border: 1px solid var(--light-grey);
-  /* color: var(--light-grey); */
-}
-
-.user-prompt .error {
-  color: red;
-  font-size: 1.1rem;
-}
-
-.user-prompt button {
-  display: flex;
-  flex-direction: row;
-  padding: 1rem 2rem;
-  border-radius: 2rem;
-  align-items: center;
-  color: white;
-  background-color: var(--light-blue);
-  line-height: 1.5rem;
-  outline: none;
-}
-
-.user-prompt hr {
-  width: 100%
-}
-
-.user-prompt .tip {
-  color: var(--light-grey);
-  width: 75%;
-  font-size: 1.1rem;
-  line-height: 1.5;
-  background-image: "/images"
-}
-
-.user-prompt button:hover {
-  box-shadow: 0px 0px 5px 0px var(--light-grey);
-  background-color: var(--dark-blue);
-}
-
-.user-prompt button .content {
-  padding-top: 0.5px; 
-}
-
-.user-prompt button.clicked {
-  background: linear-gradient(152.34deg, #3398ff 25.66%, #007dfd 78.07%);
-}
-
-.user-prompt button.clicked .img.refresh {
-  animation: spin .7s linear infinite;
-}
-
-.user-prompt .tip .record {
-  margin: 0 .5rem;
-  border: 1px solid var(--faint-red);
-  padding: 2px 5px;
-  border-radius: 5px;
-  font-size: .8rem;
-  vertical-align: text-bottom;
-  color: var(--faint-red)
-}
-
-.user-prompt-container {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  background-color: white;
-}
-
-.user-prompt-container.error {
-  background: linear-gradient(152.2deg, #67DAFF 6.75%, #4382FF 93.14%);
-  z-index: 1000;
-}
-
-.user-prompt-container.error.overlay {
-  background: linear-gradient(152.2deg, rgba(103, 218, 255, 0.90) 6.75%, rgba(67, 130, 255, 0.90) 93.14%);
-  z-index: 1000;
-}
-
-.user-prompt-container.error .user-prompt {
-  background-color: white;
-  min-height: 460px;
-  min-width: 630px;
-  border-radius: 12px;
-  justify-content: center;
-  background-image: url(https://svgshare.com/i/Q8u.svg);
-  background-repeat: no-repeat;
-  background-position: center;
-}
-
-.user-prompt button .img.refresh {
-  margin-inline-end: .5em;
-  background-color: white;
-}
